+++ src/f32x4_.rs
@@ -1177,7 +1177,7 @@ impl f32x4 {
     re = re.mul_add(zz * z, z) + s;
 
     // get sign bit
-    re = (self.sign_bit()).blend(-re, re);
+    re = (self.is_sign_negative()).blend(-re, re);
 
     re
   }
@@ -1230,10 +1230,10 @@ impl f32x4 {
     // move back in place
     re = swapxy.blend(Self::FRAC_PI_2 - re, re);
     re = ((x | y).simd_eq(Self::ZERO)).blend(Self::ZERO, re);
-    re = (x.sign_bit()).blend(Self::PI - re, re);
+    re = (x.is_sign_negative()).blend(Self::PI - re, re);
 
     // get sign bit
-    re = (y.sign_bit()).blend(-re, re);
+    re = (y.is_sign_negative()).blend(-re, re);
 
     re
   }
@@ -1529,8 +1529,18 @@ impl f32x4 {
   fn nan_pow() -> Self {
     cast::<_, f32x4>(i32x4::splat(0x7FC00000 | 0x101 & 0x003FFFFF))
   }
+
   #[inline]
-  pub fn sign_bit(self) -> Self {
+  #[must_use]
+  pub fn is_sign_positive(self) -> Self {
+    let t1 = cast::<_, i32x4>(self);
+    let t2 = t1 >> 31;
+    cast::<_, f32x4>(t2).simd_eq(f32x4::ZERO)
+  }
+
+  #[inline]
+  #[must_use]
+  pub fn is_sign_negative(self) -> Self {
     let t1 = cast::<_, i32x4>(self);
     let t2 = t1 >> 31;
     !cast::<_, f32x4>(t2).simd_eq(f32x4::ZERO)
@@ -1689,7 +1699,7 @@ impl f32x4 {
       z,
     );
 
-    let x_sign = self.sign_bit();
+    let x_sign = self.is_sign_negative();
     let z = if x_sign.any() {
       // Y into an integer
       let yi = y.simd_eq(y.round());
@@ -1844,5 +1854,12 @@ impl f32x4 {
           ] }
       }
     }
+  }
+
+  #[inline]
+  #[must_use]
+  #[deprecated(since = "1.4.0", note = "renamed to `is_sign_negative`")]
+  pub fn sign_bit(self) -> Self {
+    self.is_sign_negative()
   }
 }