+++ src/f64x2_.rs
@@ -1197,7 +1197,7 @@ impl f64x2 {
     re += s + fac;
 
     // get sign bit
-    re = (self.sign_bit()).blend(-re, re);
+    re = (self.is_sign_negative()).blend(-re, re);
 
     re
   }
@@ -1274,10 +1274,10 @@ impl f64x2 {
     // move back in place
     re = swapxy.blend(Self::FRAC_PI_2 - re, re);
     re = ((x | y).simd_eq(Self::ZERO)).blend(Self::ZERO, re);
-    re = (x.sign_bit()).blend(Self::PI - re, re);
+    re = (x.is_sign_negative()).blend(Self::PI - re, re);
 
     // get sign bit
-    re = (y.sign_bit()).blend(-re, re);
+    re = (y.is_sign_negative()).blend(-re, re);
 
     re
   }
@@ -1530,7 +1530,16 @@ impl f64x2 {
   }
 
   #[inline]
-  fn sign_bit(self) -> Self {
+  #[must_use]
+  pub fn is_sign_positive(self) -> Self {
+    let t1 = cast::<_, i64x2>(self);
+    let t2 = t1 >> 63;
+    cast::<_, f64x2>(t2).simd_eq(f64x2::ZERO)
+  }
+
+  #[inline]
+  #[must_use]
+  pub fn is_sign_negative(self) -> Self {
     let t1 = cast::<_, i64x2>(self);
     let t2 = t1 >> 63;
     !cast::<_, f64x2>(t2).simd_eq(f64x2::ZERO)
@@ -1709,7 +1718,7 @@ impl f64x2 {
       z,
     );
 
-    let x_sign = self.sign_bit();
+    let x_sign = self.is_sign_negative();
     let z = if x_sign.any() {
       // Y into an integer
       let yi = y.simd_eq(y.round());