+++ src/f64x4_.rs
@@ -1056,7 +1056,7 @@ impl f64x4 {
     re += s + fac;
 
     // get sign bit
-    re = (self.sign_bit()).blend(-re, re);
+    re = (self.is_sign_negative()).blend(-re, re);
 
     re
   }
@@ -1133,10 +1133,10 @@ impl f64x4 {
     // move back in place
     re = swapxy.blend(Self::FRAC_PI_2 - re, re);
     re = ((x | y).simd_eq(Self::ZERO)).blend(Self::ZERO, re);
-    re = (x.sign_bit()).blend(Self::PI - re, re);
+    re = (x.is_sign_negative()).blend(Self::PI - re, re);
 
     // get sign bit
-    re = (y.sign_bit()).blend(-re, re);
+    re = (y.is_sign_negative()).blend(-re, re);
 
     re
   }
@@ -1363,8 +1363,18 @@ impl f64x4 {
   fn nan_pow() -> Self {
     cast::<_, f64x4>(i64x4::splat(0x7FF8000000000000 | 0x101 << 29))
   }
+
   #[inline]
-  fn sign_bit(self) -> Self {
+  #[must_use]
+  pub fn is_sign_positive(self) -> Self {
+    let t1 = cast::<_, i64x4>(self);
+    let t2 = t1 >> 63;
+    cast::<_, f64x4>(t2).simd_eq(f64x4::ZERO)
+  }
+
+  #[inline]
+  #[must_use]
+  pub fn is_sign_negative(self) -> Self {
     let t1 = cast::<_, i64x4>(self);
     let t2 = t1 >> 63;
     !cast::<_, f64x4>(t2).simd_eq(f64x4::ZERO)
@@ -1544,7 +1554,7 @@ impl f64x4 {
       z,
     );
 
-    let x_sign = self.sign_bit();
+    let x_sign = self.is_sign_negative();
 
     let z = if x_sign.any() {
       // Y into an integer